% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addInterceptDS.R
\name{addInterceptDS}
\alias{addInterceptDS}
\title{Set intercept model}
\usage{
addInterceptDS(x.mat, intercept = FALSE)
}
\arguments{
\item{x.mat}{The name of design matrix}

\item{intercept}{The indicator to turn on (=TRUE) or off (=FALSE) the intercept model}
}
\value{
The designed matrix
}
\description{
The option to turn on or off intercept model.
}
\details{
If intercept==TRUE, the "1" column was attached on the left side of x.mat, the combination was returned. If intercept==FALSE,
x.mat was returned. In the linear regression,  \eqn{y=x \beta+b}, the non-intercept model referred to b==0. In this case, the design matrix 
and response should be z-standardized. The default was non-intercept model
}
\author{
Han Cao
}
