% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMCRDS.R
\name{calcMCRDS}
\alias{calcMCRDS}
\title{Calculate missing classification rate (MCR)}
\usage{
calcMCRDS(ws, x, y, average)
}
\arguments{
\item{ws}{The list of coefficient vectors for all servers}

\item{x}{The design matrix on a target server}

\item{y, }{The response vector on a target server}

\item{average}{The indicator to average or not the prediction probability over all coefficient vectors}
}
\value{
The MCR of each coefficient vectors (or averaged MCR)
}
\description{
Calculate missing classification rate (MCR)
}
\details{
For MTL methods, ws contained the coefficient vectors of all servers. {x, y} referred to the data pairs on the target server.  
The function calculated the MCR of {x, y} of the target server using the model ws. If average==TRUE, the averaged predicted probability 
was returned.
}
\author{
Han Cao
}
