% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMSEDS.R
\name{calcMSEDS}
\alias{calcMSEDS}
\title{Calculate mean-squared errors (MSE)}
\usage{
calcMSEDS(ws, x, y, average)
}
\arguments{
\item{ws}{The list of coefficient vectors for all servers}

\item{x}{The design matrix on a target server}

\item{y, }{The response vector on a target server}

\item{average}{The indicator to average or not the prediction scores over all coefficient vectors}
}
\value{
The MSE of each coefficient vector (or averaged MSE)
}
\description{
Calculate mean-squared errors (MSE)
}
\details{
For MTL methods, ws contained the coefficient vectors of all servers. {x, y} referred to the data pairs on the target server.  
The function calculated the MSE of {x, y} of the target server using the model ws. If average==TRUE, the averaged predicted probability 
was returned.
}
\author{
Han Cao
}
