% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMyServerDataDS.R
\name{getMyServerDataDS}
\alias{getMyServerDataDS}
\title{Get data from server}
\usage{
getMyServerDataDS(remoteKey, data)
}
\arguments{
\item{remoteKey}{A string created by the server administrator}

\item{data}{The name of the target matrix on the server}
}
\value{
The requested matrix or error message
}
\description{
Send the data matrix to the valid client
}
\details{
The "Datakey" mechanism allowed the valid client retrieve the data matrix directly into the memory of the server. The validity 
was granted by the server administrator. To achieve this, the administrator generated a key and put the local copy in the dataset 
serverDataKey.myKey(project: serverDataKey; table: myKey) where the dataset was created normally using DataSHIELD. The remote copy
of the key was hand over to the trustworthy analyst. This was determined by the server administrator
}
\author{
Han Cao
}
