% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initMatricesDS.R
\name{initMatricesDS}
\alias{initMatricesDS}
\title{Initialize variables with given random seeds}
\usage{
initMatricesDS(rank, X, seedH)
}
\arguments{
\item{rank}{The rank of data matrix in factorization}

\item{X}{The name of the data matrix for factorization}

\item{seedH}{The seed for random initialization of the shared component matrix (H)}
}
\value{
The initialized variable matrices
}
\description{
The initialization of server-side variables for method dsMTL_iNMF
}
\details{
In dsMTL_iNMF, variable matrices were initialized as drawn from the uniform distribution U(0,2). Since the heterogeneous 
component variables were updated locally, so only the seed of shared component matrix (H) were synchronized among all servers and client
}
\author{
Han Cao
}
