% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_funcVal_evalDS.R
\name{LR_funcVal_evalDS}
\alias{LR_funcVal_evalDS}
\title{Access the function value of logistic loss}
\usage{
LR_funcVal_evalDS(w, x, y)
}
\arguments{
\item{w}{The current estimate of coefficient vector}

\item{x}{The design matrix}

\item{y}{The binary label vector ({1,-1} vector)}
}
\value{
The logistic loss value
}
\description{
Calculate the function value of logistic loss of a given point
}
\details{
w was received from the client to apply on the server data pairs {x, y} and calculate the logistic loss value.
}
\author{
Han Cao
}
