% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_iter_updateDS.R
\name{LR_iter_updateDS}
\alias{LR_iter_updateDS}
\title{Access the gradient and function value of logistic loss}
\usage{
LR_iter_updateDS(w, x, y)
}
\arguments{
\item{w}{The current estimate of coefficient vector}

\item{x}{The design matrix}

\item{y}{The label vector ({1,-1} vector)}
}
\value{
The gradient and  function value of a given point
}
\description{
Calculate the gradient and  function value of logistic loss of a given point
}
\details{
The current estimate of w was sent to target server and applied on {x,y}.
}
\author{
Han Cao
}
