% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LS_funcVal_evalDS.R
\name{LS_funcVal_evalDS}
\alias{LS_funcVal_evalDS}
\title{Access the function value of least-square loss}
\usage{
LS_funcVal_evalDS(w, x, y)
}
\arguments{
\item{w}{The current estimate of coefficient vector}

\item{x}{The design matrix}

\item{y}{The label vector}
}
\value{
The function value of a given point
}
\description{
Calculate the function value of least-square loss of a given point
}
\details{
The current estimate of w was sent to target server and applied on {x,y}.
}
\author{
Han Cao
}
