% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetSubjestsDS.R
\name{subsetSubjestsDS}
\alias{subsetSubjestsDS}
\title{Subset subjects given index}
\usage{
subsetSubjestsDS(symbol, idx)
}
\arguments{
\item{symbol}{The name of the design matrix}

\item{idx}{The index of the selected subjects}
}
\value{
The subset of the design matrix
}
\description{
Subset subjects for cross-validation
}
\details{
In the k-fold in-site cross-validation, the sample in each server was randomly divided into k fold. The idx was to subset the
design matrix into the training or test sample in each of k fold cross-validation
}
\author{
Han Cao
}
