% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateHDS.R
\name{updateHDS}
\alias{updateHDS}
\title{Collect server-side information for updating H}
\usage{
updateHDS(X, H, W, Hv, Sp, lamb)
}
\arguments{
\item{X}{The name of the data matrix}

\item{H}{The current estimate of shared component H}

\item{W}{The current estimate of heterogeneous component W}

\item{Hv}{The current estimate of heterogeneous component Hv}

\item{Sp, }{The hyper-parameter for controlling the sparsity}

\item{lamb}{The hyper-parameter for balancing the amount of homogeneous and heterogeneous information}
}
\value{
The server-level information for aggregating a new estimate of H
}
\description{
Estimate the shared component (H) using the local data for method dsMTL_iNMF
}
\details{
The function was performed by each server when the client attempted to aggregate information to update H. The local variables of 
the server were used to estimate the H. These information will be aggregated on the client.
}
\author{
Han Cao
}
