% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateOtherMatsDS.R
\name{updateOtherMatsDS}
\alias{updateOtherMatsDS}
\title{Update all heterogeneous variables}
\usage{
updateOtherMatsDS(X, newH, W, Hv, lamb, Sp)
}
\arguments{
\item{X}{The name of the data matrix}

\item{newH}{The aggregated estimate of shared component H}

\item{W}{The current estimate of heterogeneous component W}

\item{Hv}{The current estimate of heterogeneous component Hv}

\item{lamb}{The hyper-parameter for balancing the amount of homogeneous and heterogeneous information}

\item{Sp, }{The hyper-parameter for controlling the sparsity}
}
\value{
The updated heterogeneous variables for eacg server
}
\description{
Update the heterogeneous variables for method dsMTL_iNMF
}
\details{
The function was performed when the server received the aggregated H, the server would update all variables in the server
}
\author{
Han Cao
}
