% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtycovDS.R
\name{xtycovDS}
\alias{xtycovDS}
\title{matrix multiplication with covars}
\usage{
xtycovDS(x, y, covar, betaCov)
}
\arguments{
\item{x}{One matrix for multiplication}

\item{y}{The other matrix for multiplication}

\item{covar}{covars}

\item{betaCov}{betaCov}
}
\value{
The result matrix
}
\description{
matrix multiplication x * t(y) covars
}
\details{
This was used to estimate the maximum  \eqn{\lambda}
}
\author{
Augusto Anguita & Han Cao
}
