% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtyDS.R
\name{xtyDS}
\alias{xtyDS}
\title{matrix multiplication}
\usage{
xtyDS(x, y)
}
\arguments{
\item{x}{One matrix for multiplication}

\item{y}{The other matrix for multiplication}
}
\value{
The result matrix
}
\description{
matrix multiplication x * t(y)
}
\details{
This was used to estimate the maximum  \eqn{\lambda}
}
\author{
Han Cao
}
